Clear PacketBuffer Content When Freed.
  -- Added new PacketBuffer::Clear() method to wipe the contents
     of buffers as they are returned to the pool.

  -- Also eliminated the integer overflow vulnerability by increasing
     the size (to uint32_t) of the rFrameLen parameter to the
     WeaveMessageLayer::DecodeMessageWithLength() method.

This change addresses CVE security vulnerability: CVE-2019-5040

+++ src/lib/core/WeaveConnection.cpp
@@ -1394,7 +1394,7 @@ void WeaveConnection::HandleDataReceived(TCPEndPoint *endPoint, PacketBuffer *da
         uint8_t*            payload;
         uint16_t            payloadLen;
         PacketBuffer*       payloadBuf = NULL;
-        uint16_t            frameLen;
+        uint32_t            frameLen;
 
         packetInfo.Clear();
         con->GetPeerAddressInfo(packetInfo);
@@ -1808,7 +1808,7 @@ void WeaveConnection::HandleBleMessageReceived(BLEEndPoint *endPoint, PacketBuff
     uint8_t *payload;
     uint16_t payloadLen;
     PacketBuffer *payloadBuf;
-    uint16_t frameLen;
+    uint32_t frameLen;
     WEAVE_ERROR err;
 
     // Initialize the message info structure.
